pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["João M <jmo.nick@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
numpy = "<2.0"
faiss-cpu = "^1.8.0.post1"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
pydantic = {extras = ["email"], version = "^2.9.0"}
pydantic-settings = "^2.5.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
tzdata = "^2024.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
